+++ spec/chef/recipes/nginx_spec.rb
@@ -1,16 +1,22 @@
 require 'chef_helper'
 
 describe 'gitlab::nginx' do
-  let(:chef_run) { ChefSpec::SoloRunner.converge('gitlab::nginx') }
+  let(:chef_runner) do
+    ChefSpec::SoloRunner.new do |node|
+      node.normal['gitlab']['nginx']['enable'] = true
+      node.normal['package']['install-dir'] = '/opt/gitlab'
+    end
+  end
+
+  let(:chef_run) do
+    chef_runner.converge('gitlab::nginx')
+  end
+
   let(:gitlab_http_config) { '/var/opt/gitlab/nginx/conf/gitlab-http.conf' }
 
   before do
     allow(Gitlab).to receive(:[]).and_call_original
-    allow(Gitlab).to receive(:[]).with('node').and_return({
-                                                            'package' => {
-                                                              'install-dir' => "/opt/gitlab"
-                                                            }
-                                                          })
+    allow(Gitlab).to receive(:[]).with('node') { chef_runner.node }
 
     # generate a random number to use as error code
     @code = rand(1000)
@@ -32,28 +38,28 @@ describe 'gitlab::nginx' do
   end
 
   it 'renders an error template when a custom error is defined' do
-    chef_run.node.normal['gitlab']['nginx']['custom_error_pages'] = @nginx_errors
-    chef_run.converge('gitlab::nginx')
+    chef_runner.node.normal['gitlab']['nginx']['custom_error_pages'] = @nginx_errors
     expect(chef_run).to render_file("/opt/gitlab/embedded/service/gitlab-rails/public/#{@code}-custom.html").with_content { |content|
       expect(content).to include("TEST MESSAGE")
     }
   end
 
   it 'creates a standard error_page entry when no custom error is defined' do
+    chef_runner.node.normal['gitlab']['nginx'].delete('custom_error_pages')
     expect(chef_run).to render_file(gitlab_http_config).with_content { |content|
       expect(content).to include("error_page 404 /404.html;")
     }
   end
 
   it 'enables the proxy_intercept_errors option when custom_error_pages is defined' do
-    chef_run.node.normal['gitlab']['nginx']['custom_error_pages'] = @nginx_errors
-    chef_run.converge('gitlab::nginx')
+    chef_runner.node.normal['gitlab']['nginx']['custom_error_pages'] = @nginx_errors
     expect(chef_run).to render_file(gitlab_http_config).with_content { |content|
       expect(content).to include("proxy_intercept_errors on")
     }
   end
 
   it 'uses the default proxy_intercept_errors option when custom_error_pages is not defined' do
+    chef_runner.node.normal['gitlab']['nginx'].delete('custom_error_pages')
     expect(chef_run).to render_file(gitlab_http_config).with_content { |content|
       expect(content).not_to include("proxy_intercept_errors")
     }